+++ src/hooks/reduxHooks/useToken.ts
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 
-import { RootState } from '../types/authTypes';
+import { RootState } from '../../types/authTypes';
 
 const useToken = (): string | null => {
   const token = useSelector<RootState, string | null>(
