+++ src/hooks/reduxHooks/useUserIdName.ts
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 
-import { RootState } from '../types/authTypes';
+import { RootState } from '../../types/authTypes';
 
 const useUserIdName = () => {
   const userId = useSelector<RootState, number | null>(
